SensorActivity:-)

+++ app/src/main/java/com/example/asessucm/Model/SensorReading.java
@@ -3,10 +3,10 @@ package com.example.asessucm.Model;
 import java.io.Serializable;
 
 public class SensorReading implements Serializable {
-    float angle;
-    float timestamp;
+    double angle;
+    double timestamp;
 
-    public SensorReading(float angle, float timestamp) {
+    public SensorReading(double angle, double timestamp) {
         this.angle = angle;
         this.timestamp = timestamp;
     }
